requirements.txt
azure-functions
azure-common
msrestazure
azure-mgmt-resource
azure-cli-core